refactor(robotiq_tsf): extract SDK-poller node into a library
No behavior change. Make the node a unit tests link against instead of
recompiling sources:

- class -> poll_data_sdk_node.hpp, main() -> poll_data_sdk_main.cpp (callback
  registration moves to the ctor; equivalent, the reader thread starts later)
- node sources build once into poll_data_sdk_lib; the executable is a thin main
- tests link the compiled lib instead of recompiling ../src/*.cpp, and use
  ament's gtest main (drop gtest_main.cpp + SKIP_LINKING_MAIN_LIBRARIES)
